Fix post tag integration test
