+++ tests/service/tags/user.rs
@@ -1,30 +1,9 @@
 use anyhow::Result;
-use serde_json::Value;
 
-use super::{make_request, TagMockup};
+use super::utils::{analyse_tag_details_structure, compare_tag_details, make_request, TagMockup};
 
 // Arst user from test.cypher
 const PUBKY_PEER: &str = "5f4e8eoogmkhqeyo5ijdix3ma6rw9byj8m36yrjp78pnxxc379to";
-
-// Small unit test to compare all the tags composition
-fn analyse_hot_tags_structure(tags: &Vec<Value>) {
-    for tag in tags {
-        assert!(tag["label"].is_string(), "label should be a string");
-        assert!(tag["taggers"].is_array(), "taggers should be an array");
-        assert!(
-            tag["taggers_count"].is_number(),
-            "taggers_count should be a number"
-        );
-    }
-}
-
-// Small unit test to compare the tag properties
-fn compare_unit_hot_tag(tag: &Value, hot_tag: TagMockup) {
-    assert_eq!(tag["label"], hot_tag.label);
-    let tagger_ids = tag["taggers"].as_array().unwrap();
-    assert_eq!(tagger_ids.len(), hot_tag.taggers);
-    assert_eq!(tag["taggers_count"], hot_tag.taggers_count);
-}
 
 #[tokio::test]
 async fn test_full_user_tags_endpoint() -> Result<()> {
@@ -37,11 +16,11 @@ async fn test_full_user_tags_endpoint() -> Result<()> {
     assert_eq!(tags.len(), 4);
 
     // Validate that the posts belong to the specified user's bookmarks
-    analyse_hot_tags_structure(tags);
+    analyse_tag_details_structure(tags);
 
     // // Analyse the tag that is in the 4th index
     let hot_tag = TagMockup::new(String::from("privacy"), 2, 2);
-    compare_unit_hot_tag(&tags[1], hot_tag);
+    compare_tag_details(&tags[1], hot_tag);
 
     Ok(())
 }
@@ -57,11 +36,11 @@ async fn test_user_tags_limit_tag_filter_active() -> Result<()> {
     assert_eq!(tags.len(), 2);
 
     // Validate that the posts belong to the specified user's bookmarks
-    analyse_hot_tags_structure(tags);
+    analyse_tag_details_structure(tags);
 
     // // Analyse the tag that is in the 4th index
     let hot_tag = TagMockup::new(String::from("pubky"), 3, 3);
-    compare_unit_hot_tag(&tags[0], hot_tag);
+    compare_tag_details(&tags[0], hot_tag);
 
     Ok(())
 }
@@ -77,11 +56,11 @@ async fn test_user_tags_limit_taggers_filter_active() -> Result<()> {
     assert_eq!(tags.len(), 4);
 
     // Validate that the posts belong to the specified user's bookmarks
-    analyse_hot_tags_structure(tags);
+    analyse_tag_details_structure(tags);
 
     // // Analyse the tag that is in the 4th index
     let hot_tag = TagMockup::new(String::from("hike"), 1, 2);
-    compare_unit_hot_tag(&tags[2], hot_tag);
+    compare_tag_details(&tags[2], hot_tag);
 
     Ok(())
 }
@@ -97,11 +76,11 @@ async fn test_user_tags_full_filter_active() -> Result<()> {
     assert_eq!(tags.len(), 1);
 
     // Validate that the posts belong to the specified user's bookmarks
-    analyse_hot_tags_structure(tags);
+    analyse_tag_details_structure(tags);
 
     // // Analyse the tag that is in the 4th index
     let hot_tag = TagMockup::new(String::from("pubky"), 1, 3);
-    compare_unit_hot_tag(&tags[0], hot_tag);
+    compare_tag_details(&tags[0], hot_tag);
 
     Ok(())
 }
